feat: CI-tested plugin examples in Python, JAX, PyTorch, Julia, Rust
All plugin examples are now built and tested by meson:

- Python/NumPy harmonic tether (0.14s)
- Python/JAX Lennard-Jones with jax.grad autodiff (0.80s)
- Python/PyTorch feed-forward NN with torch.autograd (1.49s)
- Julia harmonic spring via PackageCompiler pattern (1.05s)
- Rust Morse potential via cdylib (cargo build in meson custom_target)

Each example runs its self-test (__main__ / main()) with assertions:
energy correctness, Newton's third law, finite-difference force
verification (JAX, PyTorch).

pixi.toml: added plugintest feature with numpy, julia, jax, torch
deps and plugintest environment.

Fixes:
- JAX plugin: reshape forces before Newton's third law check
- PyTorch plugin: cast model to float64 to match input dtype
- Rust plugin: use Once + Box::into_raw instead of static descriptor
  (raw pointers aren't Sync)

+++ CppCore/examples/plugins/torch_nn_plugin.py
@@ -99,7 +99,7 @@ class TorchNNPotential:
         if model_path:
             self.model = torch.jit.load(model_path, map_location=self.device)
         else:
-            self.model = PairwiseNN(cutoff=cutoff).to(self.device)
+            self.model = PairwiseNN(cutoff=cutoff).double().to(self.device)
             self.model.eval()
 
     def calculate(
